Reorganize source_map directory
