+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/source_map/get_source_file_info.rs
@@ -1,4 +1,6 @@
-use swc_common::{source_map::Pos, BytePos, Loc, SourceMap, Span};
+use swc_common::{source_map::Pos, BytePos, SourceMap, Span};
+
+use super::private_get_source_file_info::PrivateGetSourceFileInfo;
 
 pub type Range = (usize, usize);
 
@@ -20,40 +22,6 @@ pub trait GetSourceFileInfo {
     ) -> String;
     fn generate_modified_source(&self, span: Span, replacement: &String) -> String;
     fn generate_modified_range(&self, span: Span, replacement: &String) -> Range;
-}
-
-trait PrivateGetSourceFileInfo {
-    fn get_loc(&self, span: Span) -> Loc;
-    fn get_start_col(&self, span: Span) -> usize;
-    fn get_end_col(&self, span: Span) -> usize;
-    fn get_well_formed_end_line(&self, span: Span) -> String;
-    fn get_well_formed_line(&self, span: Span) -> String;
-}
-
-impl PrivateGetSourceFileInfo for SourceMap {
-    fn get_loc(&self, span: Span) -> Loc {
-        self.lookup_char_pos(span.lo)
-    }
-
-    fn get_start_col(&self, span: Span) -> usize {
-        let loc = self.lookup_char_pos(span.lo);
-        loc.col_display
-    }
-
-    fn get_end_col(&self, span: Span) -> usize {
-        let loc = self.lookup_char_pos(span.hi);
-        loc.col_display
-    }
-
-    fn get_well_formed_end_line(&self, span: Span) -> String {
-        let line = self.get_source(span);
-        line[self.get_end_col(span)..].to_string()
-    }
-
-    fn get_well_formed_line(&self, span: Span) -> String {
-        let line = self.get_source(span);
-        line[..self.get_start_col(span)].to_string()
-    }
 }
 
 impl GetSourceFileInfo for SourceMap {
