+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/source_map/mod.rs
@@ -1,4 +1,5 @@
-pub mod source_map;
+pub use get_source_file_info::GetSourceFileInfo;
+pub use get_source_file_info::Range;
 
-pub use source_map::GetSourceFileInfo;
-pub use source_map::Range;
+pub mod get_source_file_info;
+pub mod private_get_source_file_info;
